Refactor code base (#37)
