+++ src/common/Backdrop/Backdrop.js
@@ -2,7 +2,12 @@ import React from 'react'
 import BackdropMui from '@mui/material/Backdrop';
 import CircularProgress from '@mui/material/CircularProgress';
 
-function Backdrop({backdropOpen}) {
+/** 
+ * Display circular loader with a hazed shadow background.
+ * @param {boolean} backdropOpen - Stores flag to display / hide backdrop.
+ * @return {JSX.Element} Mui backdrop component
+ */
+const Backdrop = ({backdropOpen}) => {
   return (
     <BackdropMui
         sx={{ color: '#fff', zIndex: (theme) => theme.zIndex.drawer + 1 }}
