+++ src/components/Sidebar/Row/Row.js
@@ -2,10 +2,11 @@ import React from 'react'
 
 /**
  * Row that will be displayed on each section.
- * @param {String} rowTitle - Title to be given on each row.
- * @param {String} rowIcon - Icon to be placed on each row of a section.
- * @param {String} rowParent - Parent section of a row.
- * @returns {JSX.Element} JSX row that should be shown on each section.
+ * @param {string} rowTitle - Title to be given on each row.
+ * @param {JSX.Element} rowIcon - Icon to be placed on each row of a section.
+ * @param {string} rowParent - Parent section of a row.
+ * @param {fuction} setQuery - Sets the value of the code editor.
+ * @returns {JSX.Element} Row that should be shown in each section.
  */
 
 const Row = ({ rowTitle, rowIcon, rowParent, setQuery }) => {
