+++ src/components/Sidebar/Sidebar.css
@@ -35,5 +35,5 @@
 }
 .sidebar .sidebar__section--para .section-icon{
     font-size: 14px !important;
-    margin-right: 5px;
+    margin-right: 10px;
 }
