+++ src/data/query.js
@@ -1,8 +1,11 @@
+/**
+ * Pre defined sets of query.
+*/
 const query = [ 
-    "SELECT * FROM 'products';",
-    "SELECT * FROM 'regions';",
-    "SELECT * FROM 'employees';",
-    "SELECT * FROM 'orders';",
-    "SELECT * FROM 'territories';"
+    "SELECT * FROM 'territories' WHERE 'territoryDescription' = 'NewYork';",
+    "SELECT * FROM 'orders' WHERE 'employeeID' <= 5;",
+    "SELECT * FROM 'products' WHERE 'categoryID' != 3;",
+    "SELECT * FROM 'orders' WHERE 'shipName' = 'Wartian Herkku';",
+    "SELECT * FROM 'shippers' WHERE 'shipperID' = 1;"
 ];
 export {query};
