+++ src/services/tableHttpRequest.js
@@ -1,3 +1,8 @@
+/**
+ * Parse CSV data sent from Github.
+ * @param {string} text - Stores csv raw data.
+ * @return {array} table rows json data.
+*/
 const parseCSV = (text, rowDelimiter = "\n", entryDelimiter = ",") => {
     const lines = text.split(rowDelimiter);
     const headers = lines[0].split(entryDelimiter);
@@ -13,6 +18,11 @@ const parseCSV = (text, rowDelimiter = "\n", entryDelimiter = ",") => {
     return rows;
 };
 
+/**
+ * Extract data for each column of a table rows.
+ * @param {string} fieldName - Stores key as a column name.
+ * @return {string} value for each column of a row.
+*/
 const getFieldDetails = (fieldName) => {
     if (fieldName.search("ID") !== -1) {
         return {
@@ -26,7 +36,11 @@ const getFieldDetails = (fieldName) => {
     };
 };
 
-
+/**
+ * Invoke api calls to get table rows.
+ * @param {string} tableName - Stores name of the table used to make http request.
+ * @return {array} table rows json data.
+*/
 let fetchCSVData = (tableName) =>{
     return new Promise((resolve, reject) => {
         fetch(`https://api.github.com/repos/graphql-compose/graphql-compose-examples/contents/examples/northwind/data/csv/${tableName}.csv?ref=master`)
@@ -34,24 +48,21 @@ let fetchCSVData = (tableName) =>{
         .then(
             (res) => {
                 try{
-                    // GitHub sends over base64 encoded content
-                    const rawResults = parseCSV(atob(res.content))
+                    const rawResults = parseCSV(atob(res.content));  // base64 encoded content.
                     let table = rawResults.map((rawResult) => {
                         // Use the custom processing function for each field type.
                         Object.keys(rawResult).forEach((key) => {
-                            rawResult[key] = getFieldDetails(key).processFn(
-                                rawResult[key]
-                            );
+                            rawResult[key] = getFieldDetails(key).processFn(rawResult[key]);
                         });
                         return rawResult;
                     })
                     resolve(table);
                 }catch(e){
-                    reject(`${tableName} table doesn't exist!`);
+                    reject(`No records found, try different query?`);
                 }
             },
             (err) => {
-                reject("Something went wrong!");
+                reject("Something went wrong, please try again.");
             }
         )
     });
